Webserver queue implemented and build works.
The python scripts for building, flashing and testing are written so
they can be called as standalone script or one can simply include them
in another Python script and call the respective functions instead. The
connection between the web server and the three scripts are not fully
completed. Currently there are some hard coded paths that needs to be
replaced with configuration variables in the future.

With this commit, the following oneliner will: setup everything, build
all, flash, boot and run xtest.
 $ ./hab_builder.py && sleep 1 && ./hab_flash.py && sleep 1 && ./hab_xtest.py

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>